LETS GET SOUNDIN'
Added sfx!! Steps are a WIP, but everything else works fine.
The .json  for every character now require  a new parameter "codeName",  which must be  the same as the sfx name implemented at start. Ex.: Blinkmaster hast a codeName "cat".
This is to play different sfx when doing general things.